Crate gloo_dialogs

This crate provides wrapper for alert, prompt and confirm functions. web-sys provides a raw API which is hard to use. This crate provides an easy-to-use, idiomatic Rust API for these functions.

See the documentation for alert, prompt and confirm for more information.

Functions

alert

Calls the alert function.

confirm

Calls the confirm function.

prompt

Calls the prompt function.